move global_topk to explantion
