+++ actions/get_action_functions.py
@@ -18,7 +18,7 @@ from actions.explanation.rationalize import rationalize_operation
 #from actions.explanation.what_if important what_if_operation
 from actions.filter.filter import filter_operation
 from actions.filter.includes_token import includes_operation
-from actions.global_topk import global_top_k
+from actions.explanation.global_topk import global_top_k
 from actions.metadata.count_data_points import count_data_points
 from actions.metadata.data_summary import data_operation
 from actions.metadata.feature_stats import feature_stats
